Simplify management of json-ld context and id
