+++ customer-commons/src/main/java/no/unit/nva/customer/Constants.java
@@ -1,6 +1,7 @@
 package no.unit.nva.customer;
 
 import com.google.common.net.MediaType;
+import java.net.URI;
 import no.unit.nva.customer.service.CustomerService;
 import no.unit.nva.customer.service.impl.DynamoDBCustomerService;
 import nva.commons.apigateway.MediaTypes;
@@ -13,6 +14,11 @@ import static no.unit.nva.database.DatabaseConfig.DEFAULT_DYNAMO_CLIENT;
 
 @JacocoGenerated
 public final class Constants {
+
+    public static final String LINKED_DATA_ID = "id";
+    public static final String LINKED_DATA_CONTEXT = "@context";
+    public static final URI LINKED_DATA_CONTEXT_VALUE =
+        URI.create("https://bibsysdev.github.io/src/customer-context.json");
 
     public static final List<MediaType> DEFAULT_RESPONSE_MEDIA_TYPES = List.of(
         MediaType.JSON_UTF_8,
