+++ customer-commons/src/test/java/no/unit/nva/customer/model/CustomerDtoTest.java
@@ -56,10 +56,8 @@ class CustomerDtoTest {
         return CustomerDto.builder()
                 .withCname(randomString())
                 .withIdentifier(UUID.randomUUID())
-                .withId(randomUri())
                 .withDisplayName(randomString())
                 .withInstitutionDns(randomString())
-                .withContext(randomUri())
                 .withShortName(randomString())
                 .withArchiveName(randomString())
                 .withName(randomString())
