+++ customer-commons/src/test/java/no/unit/nva/customer/model/CustomerReferenceTest.java
@@ -13,7 +13,6 @@ class CustomerReferenceTest {
     @Test
     void shouldCreateCustomerReferenceFromCustomerDto() {
         var customerDto = CustomerDto.builder()
-                              .withId(randomUri())
                               .withDisplayName(randomString())
                               .withCreatedDate(Instant.now())
                               .build();
